refactor(core): split reflect.hpp from type.hpp
